requirements.txt
python-whois
scapy-python3==0.23
requests